refactor(controller) : Lotto 리팩토링
로또의 크기인 6 상수화
toString을 통해 출력을 쉽게 함.
